requirements.txt
# Data & storage
duckdb
pandas
numpy
faker

# Monitoring & ML
scikit-learn
scipy

# Config
pyyaml

# LLM
anthropic
python-dotenv

# Alerts & lineage
requests
networkx
pyvis

# Frontend
streamlit>=1.30.0

# Testing
pytest